refactor: Reorganize folder structure for attendance feature
