+++ src/assets/attendancePage/img/flexBird.png
