+++ src/assets/attendancePage/img/poorBird.png
