+++ src/assets/attendancePage/img/richBird.png
