+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './assets/index.css';
-import ViewAllAttendance from './pages/ViewAllAttendance';
+import ViewAllAttendance from './pages/attendance/ViewAllAttendance';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
